refactor: modernize type annotations to PEP 604
Replace typing.Optional/Union with X | None and X | Y everywhere
(ruff UP007/UP045) and drop the now-unused typing imports. The vendored
core/delta_calculator is deliberately untouched — it tracks upstream
delta-calculator and gets converted there.

+++ src/instrument_cluster/ui/widgets/rpm_widget.py
@@ -121,7 +121,7 @@ class RpmWidget(Widget):
         """
         Show RPM in thousands (9000 -> 9, 10250 -> 10, etc.).
         """
-        return int(round(value * 0.001))
+        return round(value * 0.001)
 
     def _compute_value_area(self) -> pygame.Rect:
         """
